refactor: voucher page
