+++ src/pages/member/VoucherCollectionAdminPage.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { useIntl } from 'react-intl'
 import DefaultLayout from '../../components/layout/DefaultLayout'
 import MemberAdminLayout from '../../components/layout/MemberAdminLayout'
-import VoucherCollectionBlock from '../../containers/voucher/VoucherCollectionBlock'
+import VoucherCollectionBlock from '../../components/voucher/VoucherCollectionBlock'
 import { commonMessages } from '../../helpers/translation'
 import { ReactComponent as GiftIcon } from '../../images/gift.svg'
 import ForbiddenPage from '../ForbiddenPage'
