+++ src/translations/locales/en-us.json
@@ -517,7 +517,6 @@
   "common.pointsAdmin": "Points Log",
   "common.practiceManagement": "Job Management",
   "common.prepare": "Order verification takes a few seconds, your order information will be updated within a minute",
-  "common.product.activity": "活動",
   "common.product.allActivityTicket": "All Activities",
   "common.product.allAppointmentPlan": "All AppointmentPlan",
   "common.product.allMerchandise": "All Merchandise",
@@ -527,12 +526,7 @@
   "common.product.allProgramPlan": "All Program Plans",
   "common.product.allProjectPlan": "All Projects",
   "common.product.allScope": "Site-wide discount",
-  "common.product.creator": "大師",
-  "common.product.merchandise": "商品",
   "common.product.otherSpecificProduct": "Other Specific Product",
-  "common.product.podcast": "廣播",
-  "common.product.program": "線上課程",
-  "common.product.project": "專案",
   "common.profile": "View Profile",
   "common.program": "Program",
   "common.program.all": "All single courses",
@@ -584,8 +578,6 @@
   "common.text.min": "minutes",
   "common.text.noMatch": "Sorry, there is no matching match yet",
   "common.text.noOrderLog": "尚無消費紀錄",
-  "common.text.noSearchResult": "找不到相關內容",
-  "common.text.noTagContent": "找不到關於這個標籤的內容",
   "common.text.notFoundMemberEmail": "Cannot find this registered email",
   "common.text.orderSuccessHint": "If you choose \"{method}\", it will not be activated until {waitingDays} working days after payment is completed. ",
   "common.text.orderTracking": "Order Tracking",
@@ -1279,12 +1271,17 @@
   "voucher": "Exchange Voucher",
   "voucher.*.cancel": "取消",
   "voucher.*.close": "close",
+  "voucher.Voucher.fromNow": "即日起",
+  "voucher.Voucher.noUsePeriod": "無使用期限",
+  "voucher.Voucher.redeemable": "可兌換",
+  "voucher.VoucherCollectionBLock.exchangeVoucher": "兌換成功，請到「我的主頁」查看",
   "voucher.VoucherDeliverModal.copyLink": "Copy Link",
   "voucher.VoucherDeliverModal.giftLink": "Gift Link",
   "voucher.VoucherDeliverModal.transfer": "Transfer",
   "voucher.VoucherExchangeModal.emptyNoticeText": "This Product is outdated,",
   "voucher.VoucherExchangeModal.emptyNoticeText2": "please contact staff for assistance",
   "voucher.VoucherExchangeModal.exchange": "兌換",
+  "voucher.VoucherExchangeModal.exchanging": "兌換中",
   "voucher.VoucherExchangeModal.notice": "兌換券為一次使用後失效，請一次兌換完畢",
   "voucher.VoucherExchangeModal.productQuantityLimitText": "can exchange {productQuantityLimit} items",
   "voucher.VoucherExchangeModal.useNow": "立即使用",
@@ -1294,7 +1291,6 @@
   "voucher.datePicker.placeholder.startedAt": "Start Date",
   "voucher.enter": "Enter voucher code",
   "voucher.form.label.description": "Use restrictions and descriptions",
-  "voucher.fromNow": "From now on",
   "voucher.items": "items",
   "voucher.label.quantityLimit": "Quantity of redeemed items",
   "voucher.label.title": "Voucher plan name",
@@ -1303,8 +1299,6 @@
   "voucher.label.voucherCodes": "Voucher Codes",
   "voucher.modal.project.edit": "Edit Scheme",
   "voucher.modal.title.add": "Create redemption scheme",
-  "voucher.noUsePeriod": "No UsePeriod",
-  "voucher.redeemable": "redeemable",
   "voucher.tab.exchange": "Exchange Code",
   "voucher.title.add": "Add Voucher"
 }
