+++ src/translations/locales/id.json
@@ -517,7 +517,6 @@
   "common.pointsAdmin": "Log Poin",
   "common.practiceManagement": "Manajemen Pekerjaan",
   "common.prepare": "Verifikasi pesanan membutuhkan waktu beberapa detik, informasi pesanan Anda akan diperbarui dalam satu menit",
-  "common.product.activity": "活動",
   "common.product.allActivityTicket": "Semua Aktivitas",
   "common.product.allAppointmentPlan": "Semua AppointmentPlan",
   "common.product.allMerchandise": "Semua Barang Dagangan",
@@ -527,12 +526,7 @@
   "common.product.allProgramPlan": "Semua Paket Program",
   "common.product.allProjectPlan": "Semua Proyek",
   "common.product.allScope": "Diskon seluruh situs",
-  "common.product.creator": "大師",
-  "common.product.merchandise": "商品",
   "common.product.otherSpecificProduct": "Produk Spesifik Lainnya",
-  "common.product.podcast": "廣播",
-  "common.product.program": "線上課程",
-  "common.product.project": "專案",
   "common.profile": "Lihat Profil",
   "common.program": "課程",
   "common.program.all": "Semua kursus tunggal",
@@ -584,8 +578,6 @@
   "common.text.min": "menit",
   "common.text.noMatch": "Maaf, belum ada yang cocok",
   "common.text.noOrderLog": "尚無消費紀錄",
-  "common.text.noSearchResult": "找不到相關內容",
-  "common.text.noTagContent": "找不到關於這個標籤的內容",
   "common.text.notFoundMemberEmail": "Tidak dapat menemukan email terdaftar ini",
   "common.text.orderSuccessHint": "Jika Anda memilih \"{method}\", itu tidak akan diaktifkan sampai {waitingDays} hari kerja setelah pembayaran selesai. ",
   "common.text.orderTracking": "Pelacakan Pesanan",
@@ -1279,12 +1271,17 @@
   "voucher": "Voucher Tukar",
   "voucher.*.cancel": "取消",
   "voucher.*.close": "close",
+  "voucher.Voucher.fromNow": "即日起",
+  "voucher.Voucher.noUsePeriod": "無使用期限",
+  "voucher.Voucher.redeemable": "可兌換",
+  "voucher.VoucherCollectionBLock.exchangeVoucher": "兌換成功，請到「我的主頁」查看",
   "voucher.VoucherDeliverModal.copyLink": "Salin Tautan",
   "voucher.VoucherDeliverModal.giftLink": "Tautan Hadiah",
   "voucher.VoucherDeliverModal.transfer": "Transfer",
   "voucher.VoucherExchangeModal.emptyNoticeText": "This Product is outdated,",
   "voucher.VoucherExchangeModal.emptyNoticeText2": "please contact staff for assistance",
   "voucher.VoucherExchangeModal.exchange": "兌換",
+  "voucher.VoucherExchangeModal.exchanging": "兌換中",
   "voucher.VoucherExchangeModal.notice": "兌換券為一次使用後失效，請一次兌換完畢",
   "voucher.VoucherExchangeModal.productQuantityLimitText": "can exchange {productQuantityLimit} items",
   "voucher.VoucherExchangeModal.useNow": "立即使用",
@@ -1294,7 +1291,6 @@
   "voucher.datePicker.placeholder.startedAt": "Tanggal Mulai",
   "voucher.enter": "Masukkan kode voucher",
   "voucher.form.label.description": "Gunakan batasan dan deskripsi",
-  "voucher.fromNow": "Mulai sekarang",
   "voucher.items": "item",
   "voucher.label.quantityLimit": "Jumlah item yang ditukarkan",
   "voucher.label.title": "Nama paket voucher",
@@ -1303,8 +1299,6 @@
   "voucher.label.voucherCodes": "Kode Voucher",
   "voucher.modal.project.edit": "Edit Skema",
   "voucher.modal.title.add": "Buat skema penukaran",
-  "voucher.noUsePeriod": "Tidak Ada Periode Penggunaan",
-  "voucher.redeemable": "dapat ditukarkan",
   "voucher.tab.exchange": "Kode Tukar",
   "voucher.title.add": "Tambahkan Voucher"
 }
