+++ src/translations/locales/vi.json
@@ -517,7 +517,6 @@
   "common.pointsAdmin": "Nhật ký Điểm",
   "common.practiceManagement": "Quản lý Công việc",
   "common.prepare": "Quá trình xác minh đơn hàng mất vài giây, thông tin đơn hàng của bạn sẽ được cập nhật trong vòng một phút",
-  "common.product.activity": "活動",
   "common.product.allActivityTicket": "Tất cả Hoạt động",
   "common.product.allAppointmentPlan": "Tất cả lịch hẹn",
   "common.product.allMerchandise": "Tất cả hàng hóa",
@@ -527,12 +526,7 @@
   "common.product.allProgramPlan": "Tất cả các kế hoạch chương trình",
   "common.product.allProjectPlan": "Tất cả các dự án",
   "common.product.allScope": "Giảm giá trên toàn trang",
-  "common.product.creator": "大師",
-  "common.product.merchandise": "商品",
   "common.product.otherSpecificProduct": "Sản phẩm cụ thể khác",
-  "common.product.podcast": "廣播",
-  "common.product.program": "線上課程",
-  "common.product.project": "專案",
   "common.profile": "Xem hồ sơ",
   "common.program": "Chương trình",
   "common.program.all": "Tất cả các khóa học đơn lẻ",
@@ -584,8 +578,6 @@
   "common.text.min": "phút",
   "common.text.noMatch": "Xin lỗi, chưa có kết quả phù hợp nào",
   "common.text.noOrderLog": "尚無消費紀錄",
-  "common.text.noSearchResult": "找不到相關內容",
-  "common.text.noTagContent": "找不到關於這個標籤的內容",
   "common.text.notFoundMemberEmail": "Không thể tìm thấy email đã đăng ký này",
   "common.text.orderSuccessHint": "Nếu bạn chọn\" {method} \", nó sẽ không được kích hoạt cho đến {waitDays} ngày làm việc sau khi thanh toán hoàn tất. ",
   "common.text.orderTracking": "Theo dõi Đơn hàng",
@@ -1279,12 +1271,17 @@
   "voucher": "Phiếu trao đổi",
   "voucher.*.cancel": "取消",
   "voucher.*.close": "close",
+  "voucher.Voucher.fromNow": "即日起",
+  "voucher.Voucher.noUsePeriod": "無使用期限",
+  "voucher.Voucher.redeemable": "可兌換",
+  "voucher.VoucherCollectionBLock.exchangeVoucher": "兌換成功，請到「我的主頁」查看",
   "voucher.VoucherDeliverModal.copyLink": "Sao chép liên kết",
   "voucher.VoucherDeliverModal.giftLink": "Liên kết quà tặng",
   "voucher.VoucherDeliverModal.transfer": "Chuyển khoản",
   "voucher.VoucherExchangeModal.emptyNoticeText": "This Product is outdated,",
   "voucher.VoucherExchangeModal.emptyNoticeText2": "please contact staff for assistance",
   "voucher.VoucherExchangeModal.exchange": "兌換",
+  "voucher.VoucherExchangeModal.exchanging": "兌換中",
   "voucher.VoucherExchangeModal.notice": "兌換券為一次使用後失效，請一次兌換完畢",
   "voucher.VoucherExchangeModal.productQuantityLimitText": "can exchange {productQuantityLimit} items",
   "voucher.VoucherExchangeModal.useNow": "立即使用",
@@ -1294,7 +1291,6 @@
   "voucher.datePicker.placeholder.startedAt": "Ngày bắt đầu",
   "voucher.enter": "Nhập mã voucher",
   "voucher.form.label.description": "Giới hạn sử dụng và mô tả",
-  "voucher.fromNow": "Từ bây giờ",
   "voucher.items": "mặt hàng",
   "voucher.label.quantityLimit": "Số lượng các mặt hàng đã đổi",
   "voucher.label.title": "Tên gói phiếu mua hàng",
@@ -1303,8 +1299,6 @@
   "voucher.label.voucherCodes": "Mã phiếu thưởng",
   "voucher.modal.project.edit": "Chỉnh sửa sơ đồ",
   "voucher.modal.title.add": "Tạo chương trình đổi quà",
-  "voucher.noUsePeriod": "No UsePeriod",
-  "voucher.redeemable": "có thể đổi được",
   "voucher.tab.exchange": "Mã trao đổi",
   "voucher.title.add": "Thêm Voucher"
 }
