+++ src/translations/locales/zh-cn.json
@@ -517,7 +517,6 @@
   "common.pointsAdmin": "点数纪录",
   "common.practiceManagement": "作业管理",
   "common.prepare": "订单查验需要数秒的时间，将于一分钟内会更新您的订单资讯",
-  "common.product.activity": "活動",
   "common.product.allActivityTicket": "全部活动",
   "common.product.allAppointmentPlan": "全部预约",
   "common.product.allMerchandise": "全部商品",
@@ -527,12 +526,7 @@
   "common.product.allProgramPlan": "全部课程方案",
   "common.product.allProjectPlan": "全部专案",
   "common.product.allScope": "全站折抵",
-  "common.product.creator": "大師",
-  "common.product.merchandise": "商品",
   "common.product.otherSpecificProduct": "其他特定项目",
-  "common.product.podcast": "廣播",
-  "common.product.program": "線上課程",
-  "common.product.project": "專案",
   "common.profile": "查看简介",
   "common.program": "课程",
   "common.program.all": "所有单次课程",
@@ -584,8 +578,6 @@
   "common.text.min": "分钟",
   "common.text.noMatch": "很抱歉，目前还没有对应的组合",
   "common.text.noOrderLog": "尚無消費紀錄",
-  "common.text.noSearchResult": "找不到相關內容",
-  "common.text.noTagContent": "找不到關於這個標籤的內容",
   "common.text.notFoundMemberEmail": "找不到这个注册信箱",
   "common.text.orderSuccessHint": "若你选择「{method}」需于付款完成后，{waitingDays} 个工作日才会开通。 ",
   "common.text.orderTracking": "订单查询",
@@ -1279,12 +1271,17 @@
   "voucher": "兑换券",
   "voucher.*.cancel": "取消",
   "voucher.*.close": "close",
+  "voucher.Voucher.fromNow": "即日起",
+  "voucher.Voucher.noUsePeriod": "無使用期限",
+  "voucher.Voucher.redeemable": "可兌換",
+  "voucher.VoucherCollectionBLock.exchangeVoucher": "兌換成功，請到「我的主頁」查看",
   "voucher.VoucherDeliverModal.copyLink": "复制连结",
   "voucher.VoucherDeliverModal.giftLink": "赠送连结",
   "voucher.VoucherDeliverModal.transfer": "转赠",
   "voucher.VoucherExchangeModal.emptyNoticeText": "This Product is outdated,",
   "voucher.VoucherExchangeModal.emptyNoticeText2": "please contact staff for assistance",
   "voucher.VoucherExchangeModal.exchange": "兌換",
+  "voucher.VoucherExchangeModal.exchanging": "兌換中",
   "voucher.VoucherExchangeModal.notice": "兌換券為一次使用後失效，請一次兌換完畢",
   "voucher.VoucherExchangeModal.productQuantityLimitText": "can exchange {productQuantityLimit} items",
   "voucher.VoucherExchangeModal.useNow": "立即使用",
@@ -1294,7 +1291,6 @@
   "voucher.datePicker.placeholder.startedAt": "开始日期",
   "voucher.enter": "输入兑换码",
   "voucher.form.label.description": "使用限制和描述",
-  "voucher.fromNow": "即日起",
   "voucher.items": "个项目",
   "voucher.label.quantityLimit": "兑换项目数量",
   "voucher.label.title": "兑换方案名称",
@@ -1303,8 +1299,6 @@
   "voucher.label.voucherCodes": "兑换码",
   "voucher.modal.project.edit": "编辑方案",
   "voucher.modal.title.add": "建立兑换方案",
-  "voucher.noUsePeriod": "无使用期限",
-  "voucher.redeemable": "可兑换",
   "voucher.tab.exchange": "兑换代码",
   "voucher.title.add": "新增兑换券"
 }
