+++ src/translations/locales/zh-tw.json
@@ -517,7 +517,6 @@
   "common.pointsAdmin": "點數紀錄",
   "common.practiceManagement": "作業管理",
   "common.prepare": "訂單查驗需要數秒的時間，將於一分鐘內會更新您的訂單資訊",
-  "common.product.activity": "活動",
   "common.product.allActivityTicket": "全部活動",
   "common.product.allAppointmentPlan": "全部預約",
   "common.product.allMerchandise": "全部商品",
@@ -527,12 +526,7 @@
   "common.product.allProgramPlan": "全部課程方案",
   "common.product.allProjectPlan": "全部專案",
   "common.product.allScope": "全站折抵",
-  "common.product.creator": "大師",
-  "common.product.merchandise": "商品",
   "common.product.otherSpecificProduct": "其他特定項目",
-  "common.product.podcast": "廣播",
-  "common.product.program": "線上課程",
-  "common.product.project": "專案",
   "common.profile": "查看簡介",
   "common.program": "課程",
   "common.program.all": "所有單次課程",
@@ -584,8 +578,6 @@
   "common.text.min": "分鐘",
   "common.text.noMatch": "很抱歉，目前還沒有對應的組合",
   "common.text.noOrderLog": "尚無消費紀錄",
-  "common.text.noSearchResult": "找不到相關內容",
-  "common.text.noTagContent": "找不到關於這個標籤的內容",
   "common.text.notFoundMemberEmail": "找不到這個註冊信箱",
   "common.text.orderSuccessHint": "若你選擇「{method}」需於付款完成後，{waitingDays} 個工作日才會開通。",
   "common.text.orderTracking": "訂單查詢",
@@ -1279,12 +1271,17 @@
   "voucher": "兌換券",
   "voucher.*.cancel": "取消",
   "voucher.*.close": "關閉",
+  "voucher.Voucher.fromNow": "即日起",
+  "voucher.Voucher.noUsePeriod": "無使用期限",
+  "voucher.Voucher.redeemable": "可兌換",
+  "voucher.VoucherCollectionBLock.exchangeVoucher": "兌換成功，請到「我的主頁」查看",
   "voucher.VoucherDeliverModal.copyLink": "複製連結",
   "voucher.VoucherDeliverModal.giftLink": "贈送連結",
   "voucher.VoucherDeliverModal.transfer": "轉贈",
   "voucher.VoucherExchangeModal.emptyNoticeText": "可兌換的項目已過期",
   "voucher.VoucherExchangeModal.emptyNoticeText2": "請洽客服人員",
   "voucher.VoucherExchangeModal.exchange": "兌換",
+  "voucher.VoucherExchangeModal.exchanging": "兌換中",
   "voucher.VoucherExchangeModal.notice": "兌換券為一次使用後失效，請一次兌換完畢",
   "voucher.VoucherExchangeModal.productQuantityLimitText": "可兌換 {productQuantityLimit} 個項目",
   "voucher.VoucherExchangeModal.useNow": "立即使用",
@@ -1294,7 +1291,6 @@
   "voucher.datePicker.placeholder.startedAt": "開始日期",
   "voucher.enter": "輸入兌換碼",
   "voucher.form.label.description": "使用限制和描述",
-  "voucher.fromNow": "即日起",
   "voucher.items": "個項目",
   "voucher.label.quantityLimit": "兌換項目數量",
   "voucher.label.title": "兌換方案名稱",
@@ -1303,8 +1299,6 @@
   "voucher.label.voucherCodes": "兌換碼",
   "voucher.modal.project.edit": "編輯方案",
   "voucher.modal.title.add": "建立兌換方案",
-  "voucher.noUsePeriod": "無使用期限",
-  "voucher.redeemable": "可兌換",
   "voucher.tab.exchange": "兌換代碼",
   "voucher.title.add": "新增兌換券"
 }
